[refactor]: 불필요 useMutation, useQuery로 로직 변환, 중복 로직 함수화, onChangeUserSession 조건부 처리 수정
